Tidy up: deletes empty example folder left by mistake
Hi all, thanks for sharing this great module.  I see this project is in sunset, but I found it to be the only way to sort out how GCP serves static assets without relying on Cloud Run which may be slower (and slightly more costly in high traffic scenarios?).  I'm just leaving this remark in case a maintainer materializes or I wind up frequenting enough over the next year to email y'all about maintaining.  

It looks like this README.md was left by mistake when the `http-load-balancer-website` folder seems to have taken over as the example to get started with this module.  I believe removing this folder will help new-comers focus on the existing example with one less distraction in the way.